requirements.txt
docling
fastapi>=0.116.1
fastembed>=0.7.1
pydantic>=2.11.7
pydantic-ai-slim[openai,logfire]
qdrant-client>=1.14.3
uvicorn>=0.35.0
googlesearch-python>=1.3.0
beautifulsoup4>=4.13.4
httpx>=0.28.1